re #151 rewrote a lot of things giving all viewmodels a proper cleanup.
Dispatcher is now access via the App object.
Starting and stopping the network and the realtime data is now done via
the App object, a single place.
All database calls have their tasks put in a queue in app that is used to
make sure the database calls and REST calls are complete before killing
things.